fix: rename file and content of in-outs pages
