+++ content/lessons/chapter-6/in-out-1.tsx
@@ -1,38 +1,34 @@
 'use client'
 
-import { useTranslations, useSaveAndProceed } from 'hooks'
-import { ChapterIntro, CodeExample } from 'ui'
-import { Button } from 'shared'
+import { useTranslations } from 'hooks'
+import { Introduction, Text } from 'ui'
 
 export const metadata = {
   title: 'chapter_six.in_out_one.title',
+  image: '/assets/images/chapter-4-intro-1.jpg',
   key: 'CH6INO1',
 }
 
-export default function IntOut1({ lang }) {
-  const saveAndProceed = useSaveAndProceed()
+export default function InOut3({ lang }) {
   const t = useTranslations(lang)
 
   return (
-    <ChapterIntro className="m-auto my-8 max-w-[840px]">
-      <p className="mt-2 text-lg md:text-xl">
-        {t('chapter_six.in_out_one.paragraph_one')}
-      </p>
-      <CodeExample
-        className="mt-4 "
-        code={`bc1qgghq08syehkym52ueu9nl5x8gth23vr8hurv9dyfcmhaqk4lrlgs28epwj`}
-        language="shell"
-      />
-      <p className="mt-8 text-lg md:text-xl">
-        {t('chapter_six.in_out_one.paragraph_two')}
-      </p>
-
-      <p className="mt-8 text-lg md:text-xl">
-        {t('chapter_six.in_out_one.paragraph_three')}
-      </p>
-      <Button onClick={saveAndProceed} classes="mt-10 max-md:w-full">
-        {t('shared.next')}
-      </Button>
-    </ChapterIntro>
+    <Introduction lang={lang} imagePosition="object-[50%_27%]">
+      <Text className="text-lg md:text-xl">
+        {t('chapter_five.intro_one.paragraph_one')}
+      </Text>
+      <Text className="mt-4 text-lg md:text-xl">
+        {t('chapter_five.intro_one.paragraph_two')}
+      </Text>
+      <Text className="mt-4 text-lg md:text-xl">
+        {t('chapter_five.intro_one.paragraph_three')}
+      </Text>
+      <Text className="mt-4 text-lg md:text-xl">
+        {t('chapter_five.intro_one.paragraph_four')}
+      </Text>
+      <Text className="mt-4 text-lg md:text-xl">
+        {t('chapter_five.intro_one.paragraph_five')}
+      </Text>
+    </Introduction>
   )
 }
