+++ content/lessons/chapter-6/in-out-3.tsx
@@ -4,30 +4,24 @@ import { useTranslations } from 'hooks'
 import { Introduction, Text } from 'ui'
 
 export const metadata = {
-  title: 'chapter_six.in_out_one.title',
-  image: '/assets/images/chapter-4-intro-1.jpg',
+  title: 'chapter_six.in_out_two.title',
+  image: '/assets/images/chapter-6-in-out-2.jpg',
   key: 'CH6INO3',
 }
 
-export default function InOut3({ lang }) {
+export default function InOut2({ lang }) {
   const t = useTranslations(lang)
 
   return (
     <Introduction lang={lang} imagePosition="object-[50%_27%]">
       <Text className="text-lg md:text-xl">
-        {t('chapter_five.intro_one.paragraph_one')}
+        {t('chapter_six.in_out_three.paragraph_one')}
       </Text>
       <Text className="mt-4 text-lg md:text-xl">
-        {t('chapter_five.intro_one.paragraph_two')}
+        {t('chapter_six.in_out_three.paragraph_two')}
       </Text>
       <Text className="mt-4 text-lg md:text-xl">
-        {t('chapter_five.intro_one.paragraph_three')}
-      </Text>
-      <Text className="mt-4 text-lg md:text-xl">
-        {t('chapter_five.intro_one.paragraph_four')}
-      </Text>
-      <Text className="mt-4 text-lg md:text-xl">
-        {t('chapter_five.intro_one.paragraph_five')}
+        {t('chapter_six.in_out_three.paragraph_three')}
       </Text>
     </Introduction>
   )
